Moved to mangadex api v2 and cleaned up code
Moved the title downloader into one file with chapter downloads. Group and user downloads now use the api. Title, group and user downloads now save a json of the downloaded chapter after every chapter and if the download is restarted, the downloaded chapters in the json are skipped.
Beta feature: The file checks for updates and if there are missing required files, they're downloaded.