v2.8.3 DIfferent chapter prefixes for volumes' count that start from 1 each time
Add option to save the chapter data to the downloaded archive/folder.
Remove the getLangIso() function from exporter.py
Each series has their chapters assigned to the respective volume numbers which is then checked to see if the volumes have overlapping chapter numbers that are assigned a different prefix so the downloaded files is not out-of-order.